Move iter type to own file
